test(storage): give the index-sweep counts their own suite
The sweep-count tests shared a file with the general collection tests, and most
of them restated rules that are pinned directly against the arming structure at
a fraction of the cost. What is left asks only what a collection cycle can
answer on its own: that the arming it sweeps with is claimed and reset each
cycle, that object deletion still forces the whole family, that the one arming
reaches every callee of the sweep, and that narrowing never drops an entry the
sweep was obliged to collect.

The metrics fixture moves to a header so both files can build on it.

The per-cycle sweep count stays an absolute number. It is the only probe that
tells an index swept and found empty from one never visited, which is the
distinction the arming exists to create. It counts the six index families the
sweep covers, so a seventh joining without arming support breaks the suite
rather than passing quietly.

Helpers that assert are called so a failed setup stops the test, rather than
letting the body run on against a database that was never prepared.
