+++ tests/unit/storage_v2_gc.cpp
@@ -23,71 +23,12 @@
 #include "metrics/prometheus_metrics.hpp"
 #include "storage/v2/gc_status.hpp"
 #include "storage/v2/inmemory/storage.hpp"
+#include "storage_v2_gc_metrics_fixture.hpp"
 #include "tests/test_commit_args_helper.hpp"
 
 using testing::UnorderedElementsAre;
 
 namespace ms = memgraph::storage;
-
-// Forces a synchronous GC pass by handing FreeMemory a hold it must adopt rather than acquire.
-inline auto UniqueGuard(memgraph::utils::ResourceLock &lock) {
-  return memgraph::utils::ResourceLockGuard{lock, memgraph::utils::ResourceLockGuard::UNIQUE};
-}
-
-class StorageV2GcMetricsTest : public testing::Test {
- protected:
-  void SetUp() override {
-    FLAGS_metrics_format = "OpenMetrics";
-    db_name_ = testing::UnitTest::GetInstance()->current_test_info()->name();
-    InitStorage(std::chrono::seconds(3600));
-  }
-
-  void TearDown() override {
-    memgraph::metrics::Metrics().SetStorageSnapshotResolver({});
-    storage.reset();
-    memgraph::metrics::Metrics().RemoveDatabase(uuid_);
-    handles_ = {};
-    uuid_ = {};
-    registered_ = false;
-  }
-
-  void InitStorage(std::chrono::milliseconds interval) {
-    if (registered_) {
-      memgraph::metrics::Metrics().SetStorageSnapshotResolver({});
-      storage.reset();
-      memgraph::metrics::Metrics().RemoveDatabase(uuid_);
-      handles_ = {};
-      uuid_ = {};
-      registered_ = false;
-    }
-    memgraph::storage::Config config;
-    config.salient.name = db_name_;
-    config.gc = {.type = memgraph::storage::Config::Gc::Type::PERIODIC, .interval = interval};
-    uuid_ = memgraph::utils::UUID{};
-    handles_ = memgraph::metrics::Metrics().AddDatabase(uuid_, db_name_);
-    registered_ = true;
-    storage = std::make_unique<memgraph::storage::InMemoryStorage>(
-        config, std::nullopt, std::make_unique<memgraph::storage::PlanInvalidatorDefault>(), handles_);
-    memgraph::metrics::Metrics().SetStorageSnapshotResolver(
-        [this](memgraph::utils::UUID const &uuid) -> std::optional<memgraph::metrics::StorageSnapshot> {
-          if (uuid != uuid_ || !storage) return std::nullopt;
-          auto const info = storage->GetBaseInfo();
-          return memgraph::metrics::StorageSnapshot{
-              .vertex_count = info.vertex_count,
-              .edge_count = info.edge_count,
-              .disk_usage = info.disk_usage,
-          };
-        });
-  }
-
-  std::unique_ptr<memgraph::storage::Storage> storage;
-  memgraph::metrics::DatabaseMetricHandles handles_{};
-  memgraph::utils::UUID uuid_{};
-  bool registered_{false};
-
- private:
-  std::string db_name_;
-};
 
 TEST(StorageV2GcStatus, PhaseToString) {
   using ms::GcPhase;
@@ -274,382 +215,6 @@ TEST(StorageV2Gc, Sanity) {
 // 4. Wait for GC. GC shouldn't remove the vertices from index because
 //    transaction 1 can still see them with that label.
 // NOLINTNEXTLINE(hicpp-special-member-functions)
-// The index-cleanup sweep visits every entry of an index rather than only the stale ones, so
-// what a collection cycle costs is governed by how many indexes it visits. Nothing else reports
-// that: the sweep latency says how long it took, not how much of it was worth doing.
-class StorageV2GcIndexSweepCountTest : public StorageV2GcMetricsTest {
- protected:
-  // A pass that adopts a hold it is handed, so it runs here rather than on the collection
-  // thread, and the count it produces belongs to a known set of writes.
-  uint64_t SweptByOnePass() {
-    auto const before = handles_.gc_index_sweeps.Value();
-    auto *mem_storage = static_cast<memgraph::storage::InMemoryStorage *>(storage.get());
-    mem_storage->FreeMemory(UniqueGuard(storage->main_lock_), false);
-    return static_cast<uint64_t>(handles_.gc_index_sweeps.Value() - before);
-  }
-
-  void CreateLabelIndex(std::string_view name) {
-    auto acc = storage->UniqueAccess();
-    ASSERT_TRUE(acc->CreateIndex(storage->NameToLabel(name)).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-
-  void CreateLabelPropertyIndex(std::string_view label, std::string_view property) {
-    auto acc = storage->UniqueAccess();
-    ASSERT_TRUE(acc->CreateIndex(storage->NameToLabel(label),
-                                 {memgraph::storage::PropertyPath{storage->NameToProperty(property)}})
-                    .has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-};
-
-TEST_F(StorageV2GcIndexSweepCountTest, IdleDatabaseSweepsNothing) {
-  CreateLabelIndex("A");
-  CreateLabelIndex("B");
-
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto vertex = acc->CreateVertex();
-    ASSERT_TRUE(*vertex.AddLabel(acc->NameToLabel("A")));
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-
-  // The write above says the vertex indexes may hold something to collect, so this pass sweeps.
-  EXPECT_GT(SweptByOnePass(), 0);
-
-  // Nothing has been written since, so there is nothing to look for.
-  EXPECT_EQ(SweptByOnePass(), 0);
-}
-
-TEST_F(StorageV2GcIndexSweepCountTest, OnlyTheWrittenLabelsIndexIsSwept) {
-  CreateLabelIndex("A");
-  CreateLabelIndex("B");
-
-  memgraph::storage::Gid gid;
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto vertex = acc->CreateVertex();
-    gid = vertex.Gid();
-    ASSERT_TRUE(*vertex.AddLabel(acc->NameToLabel("A")));
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  ASSERT_GT(SweptByOnePass(), 0);
-
-  // A write naming one label only. The index on the other label cannot have gained anything to
-  // collect, and walking it would cost its whole size to find that out.
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto vertex = acc->FindVertex(gid, memgraph::storage::View::OLD);
-    ASSERT_TRUE(vertex.has_value());
-    ASSERT_TRUE(*vertex->AddLabel(acc->NameToLabel("B")));
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  EXPECT_EQ(SweptByOnePass(), 1);
-}
-
-// A deleted vertex leaves index entries pointing at memory about to be freed, and no delta says
-// which indexes hold them. Narrowing must not reach this case.
-TEST_F(StorageV2GcIndexSweepCountTest, ADeletedVertexSweepsEveryVertexIndex) {
-  CreateLabelIndex("A");
-  CreateLabelIndex("B");
-
-  memgraph::storage::Gid gid;
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto vertex = acc->CreateVertex();
-    gid = vertex.Gid();
-    ASSERT_TRUE(*vertex.AddLabel(acc->NameToLabel("A")));
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  ASSERT_GT(SweptByOnePass(), 0);
-
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto vertex = acc->FindVertex(gid, memgraph::storage::View::OLD);
-    ASSERT_TRUE(vertex.has_value());
-    ASSERT_TRUE(acc->DeleteVertex(&*vertex).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  EXPECT_EQ(SweptByOnePass(), 2);
-}
-
-// The saving must not come at the cost of leaving entries behind: an index whose label was
-// written is swept, and collects everything a sweep of it would.
-TEST_F(StorageV2GcIndexSweepCountTest, TheSweptIndexStillCollectsItsStaleEntries) {
-  CreateLabelIndex("A");
-  CreateLabelIndex("B");
-
-  constexpr int kVertices = 100;
-  std::vector<memgraph::storage::Gid> gids;
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    for (int i = 0; i != kVertices; ++i) {
-      auto vertex = acc->CreateVertex();
-      gids.push_back(vertex.Gid());
-      ASSERT_TRUE(*vertex.AddLabel(acc->NameToLabel("A")));
-      ASSERT_TRUE(*vertex.AddLabel(acc->NameToLabel("B")));
-    }
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  ASSERT_GT(SweptByOnePass(), 0);
-
-  auto const indexed_count = [&](std::string_view name) {
-    auto acc = storage->Access(memgraph::storage::READ);
-    auto const count = acc->ApproximateVertexCount(storage->NameToLabel(name));
-    acc->Abort();
-    return count;
-  };
-  ASSERT_EQ(indexed_count("A"), kVertices);
-  ASSERT_EQ(indexed_count("B"), kVertices);
-
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    for (auto const gid : gids) {
-      auto vertex = acc->FindVertex(gid, memgraph::storage::View::OLD);
-      ASSERT_TRUE(vertex.has_value());
-      ASSERT_TRUE(*vertex->RemoveLabel(acc->NameToLabel("A")));
-    }
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-
-  EXPECT_EQ(SweptByOnePass(), 1);
-  EXPECT_EQ(indexed_count("A"), 0);
-  EXPECT_EQ(indexed_count("B"), kVertices);
-}
-
-TEST_F(StorageV2GcIndexSweepCountTest, OnlyTheWrittenPropertysIndexIsSwept) {
-  CreateLabelPropertyIndex("L", "a");
-  CreateLabelPropertyIndex("L", "b");
-
-  memgraph::storage::Gid gid;
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto vertex = acc->CreateVertex();
-    gid = vertex.Gid();
-    ASSERT_TRUE(*vertex.AddLabel(acc->NameToLabel("L")));
-    ASSERT_TRUE(vertex.SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{1}).has_value());
-    ASSERT_TRUE(vertex.SetProperty(acc->NameToProperty("b"), memgraph::storage::PropertyValue{1}).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  ASSERT_GT(SweptByOnePass(), 0);
-
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto vertex = acc->FindVertex(gid, memgraph::storage::View::OLD);
-    ASSERT_TRUE(vertex.has_value());
-    ASSERT_TRUE(vertex->SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{2}).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  EXPECT_EQ(SweptByOnePass(), 1);
-}
-
-// A property write says the vertex indexes may hold something to collect but a delta cannot say
-// which property it was without being read for it. Were the write to arm the family and name no
-// property, every label-property index would be skipped and the entries it left would stay.
-TEST_F(StorageV2GcIndexSweepCountTest, APropertyWriteStillCollectsTheEntriesItStaled) {
-  CreateLabelPropertyIndex("L", "a");
-  CreateLabelPropertyIndex("L", "b");
-
-  constexpr int kVertices = 100;
-  std::vector<memgraph::storage::Gid> gids;
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    for (int i = 0; i != kVertices; ++i) {
-      auto vertex = acc->CreateVertex();
-      gids.push_back(vertex.Gid());
-      ASSERT_TRUE(*vertex.AddLabel(acc->NameToLabel("L")));
-      ASSERT_TRUE(vertex.SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{i}).has_value());
-      ASSERT_TRUE(vertex.SetProperty(acc->NameToProperty("b"), memgraph::storage::PropertyValue{i}).has_value());
-    }
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  ASSERT_GT(SweptByOnePass(), 0);
-
-  auto const indexed_count = [&](std::string_view property) {
-    auto acc = storage->Access(memgraph::storage::READ);
-    auto const count = acc->ApproximateVertexCount(
-        storage->NameToLabel("L"), std::array{memgraph::storage::PropertyPath{storage->NameToProperty(property)}});
-    acc->Abort();
-    return count;
-  };
-  ASSERT_EQ(indexed_count("a"), kVertices);
-
-  // Rewriting a property leaves the entry holding the old value behind for the sweep to find.
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    for (auto const gid : gids) {
-      auto vertex = acc->FindVertex(gid, memgraph::storage::View::OLD);
-      ASSERT_TRUE(vertex.has_value());
-      ASSERT_TRUE(vertex->SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{-1}).has_value());
-    }
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  ASSERT_EQ(indexed_count("a"), 2 * kVertices);
-
-  EXPECT_EQ(SweptByOnePass(), 1);
-  EXPECT_EQ(indexed_count("a"), kVertices);
-  EXPECT_EQ(indexed_count("b"), kVertices);
-}
-
-// A unique constraint keeps a skiplist keyed the way an index is and is swept the same way, so a
-// write that cannot have staled it should not cost its whole size to walk either.
-TEST_F(StorageV2GcIndexSweepCountTest, OnlyTheWrittenPropertysConstraintIsSwept) {
-  auto const create_constraint = [&](std::string_view property) {
-    auto acc = storage->UniqueAccess();
-    auto const result = acc->CreateUniqueConstraint(storage->NameToLabel("L"), {storage->NameToProperty(property)});
-    ASSERT_TRUE(result.has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  };
-  create_constraint("a");
-  create_constraint("b");
-
-  memgraph::storage::Gid gid;
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto vertex = acc->CreateVertex();
-    gid = vertex.Gid();
-    ASSERT_TRUE(*vertex.AddLabel(acc->NameToLabel("L")));
-    ASSERT_TRUE(vertex.SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{1}).has_value());
-    ASSERT_TRUE(vertex.SetProperty(acc->NameToProperty("b"), memgraph::storage::PropertyValue{1}).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  ASSERT_GT(SweptByOnePass(), 0);
-
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto vertex = acc->FindVertex(gid, memgraph::storage::View::OLD);
-    ASSERT_TRUE(vertex.has_value());
-    ASSERT_TRUE(vertex->SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{2}).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  EXPECT_EQ(SweptByOnePass(), 1);
-
-  // The constraint left unswept still holds: its entries were never stale, only unvisited.
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto other = acc->CreateVertex();
-    ASSERT_TRUE(*other.AddLabel(acc->NameToLabel("L")));
-    ASSERT_TRUE(other.SetProperty(acc->NameToProperty("b"), memgraph::storage::PropertyValue{1}).has_value());
-    EXPECT_FALSE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-}
-
-TEST_F(StorageV2GcIndexSweepCountTest, OnlyTheWrittenPropertysEdgeIndexIsSwept) {
-  auto const create_edge_index = [&](std::string_view property) {
-    auto acc = storage->ReadOnlyAccess();
-    ASSERT_TRUE(acc->CreateIndex(storage->NameToEdgeType("E"), storage->NameToProperty(property)).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  };
-  create_edge_index("a");
-  create_edge_index("b");
-
-  memgraph::storage::Gid edge_gid;
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto from = acc->CreateVertex();
-    auto to = acc->CreateVertex();
-    auto edge = acc->CreateEdge(&from, &to, acc->NameToEdgeType("E"));
-    ASSERT_TRUE(edge.has_value());
-    edge_gid = edge->Gid();
-    ASSERT_TRUE(edge->SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{1}).has_value());
-    ASSERT_TRUE(edge->SetProperty(acc->NameToProperty("b"), memgraph::storage::PropertyValue{1}).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  ASSERT_GT(SweptByOnePass(), 0);
-
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto edge = acc->FindEdge(edge_gid, memgraph::storage::View::OLD);
-    ASSERT_TRUE(edge.has_value());
-    ASSERT_TRUE(edge->SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{2}).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  EXPECT_EQ(SweptByOnePass(), 1);
-}
-
-// A removed edge leaves entries pointing at memory about to be freed, and its delta carries the
-// edge type rather than the property any index is keyed on, so nothing names them.
-TEST_F(StorageV2GcIndexSweepCountTest, ARemovedEdgeSweepsEveryEdgeIndex) {
-  auto const create_edge_index = [&](std::string_view property) {
-    auto acc = storage->ReadOnlyAccess();
-    ASSERT_TRUE(acc->CreateIndex(storage->NameToEdgeType("E"), storage->NameToProperty(property)).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  };
-  create_edge_index("a");
-  create_edge_index("b");
-
-  memgraph::storage::Gid edge_gid;
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto from = acc->CreateVertex();
-    auto to = acc->CreateVertex();
-    auto edge = acc->CreateEdge(&from, &to, acc->NameToEdgeType("E"));
-    ASSERT_TRUE(edge.has_value());
-    edge_gid = edge->Gid();
-    ASSERT_TRUE(edge->SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{1}).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  ASSERT_GT(SweptByOnePass(), 0);
-
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto edge = acc->FindEdge(edge_gid, memgraph::storage::View::OLD);
-    ASSERT_TRUE(edge.has_value());
-    ASSERT_TRUE(acc->DeleteEdge(&*edge).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  EXPECT_EQ(SweptByOnePass(), 2);
-}
-
-// An index on an edge type alone cannot be staled by writing a property: the entry holds no
-// property, and the type it does hold cannot change.
-TEST_F(StorageV2GcIndexSweepCountTest, AnEdgePropertyWriteSweepsNoEdgeTypeIndex) {
-  auto const create_edge_type_index = [&](std::string_view edge_type) {
-    auto acc = storage->ReadOnlyAccess();
-    ASSERT_TRUE(acc->CreateIndex(storage->NameToEdgeType(edge_type)).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  };
-  create_edge_type_index("E");
-  create_edge_type_index("F");
-
-  memgraph::storage::Gid from_gid, to_gid, edge_gid;
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto from = acc->CreateVertex();
-    auto to = acc->CreateVertex();
-    from_gid = from.Gid();
-    to_gid = to.Gid();
-    auto edge = acc->CreateEdge(&from, &to, acc->NameToEdgeType("E"));
-    ASSERT_TRUE(edge.has_value());
-    edge_gid = edge->Gid();
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  // Creating the edge is a structural change, so both edge-type indexes are swept.
-  EXPECT_EQ(SweptByOnePass(), 2);
-
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto edge = acc->FindEdge(edge_gid, memgraph::storage::View::OLD);
-    ASSERT_TRUE(edge.has_value());
-    ASSERT_TRUE(edge->SetProperty(acc->NameToProperty("a"), memgraph::storage::PropertyValue{1}).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  EXPECT_EQ(SweptByOnePass(), 0);
-
-  {
-    auto acc = storage->Access(memgraph::storage::WRITE);
-    auto edge = acc->FindEdge(edge_gid, memgraph::storage::View::OLD);
-    ASSERT_TRUE(edge.has_value());
-    ASSERT_TRUE(acc->DeleteEdge(&*edge).has_value());
-    ASSERT_TRUE(acc->PrepareForCommitPhase(memgraph::tests::MakeMainCommitArgs()).has_value());
-  }
-  EXPECT_EQ(SweptByOnePass(), 2);
-
-  auto acc = storage->Access(memgraph::storage::READ);
-  EXPECT_EQ(acc->ApproximateEdgeCount(acc->NameToEdgeType("E")), 0);
-  acc->Abort();
-}
-
 TEST(StorageV2Gc, Indices) {
   std::unique_ptr<memgraph::storage::Storage> storage(
       std::make_unique<memgraph::storage::InMemoryStorage>(memgraph::storage::Config{
